router to map POST to list action
A new router is needed in this case to override DRF defaults
where GET is used to output the details of an item or a list of items
and POST is used to create a new item.
The new router restricts the methods to POST only and maps
to the default list method.