DRY individual renderers for batch query
To leverage DRF's defaults, this commit extends a ViewSet and defines:
- a queryset through the `get_queryset` method which allows to dynamically
define the queryset used in the ViewSet,
here by filtering on the ids passed as a parameter.
- content_negociation_class which is responsible for picking the right renderer
to a given request, here we need to base this on the POST request body
instead of the defaults (`accept-type` or format parameter in url)
serializer, pagination and renderer classes are copied over from
IndividualViewSet.