private search results as values from 1 field
This new feature allows to reduce the response from a free-form query to
the values from a given field.
The use case is to return the list of values for a linked field when
performing aggregations accross services: for
example querying the biosamples_id corresponding to phenopackets matching
a query for FEMALE individuals.

This commit contains refactored code to DRY duplicates between
search and chord_table_search functions.

Note: the "output" parameter was named "format" initially but it broke
down the requests when run in GET method,
as DRF kicks-in to interpret the format parameter and returns a cryptic
{details: "Not found"} error content.