Fix: pagination clamps CSV content
The paginator from DRF extracts page_size from the GET
request.
In the former implementation, the client code was
requesting for a predefined number of rows, set to
the number of ids requested.
To avoid adding this parameter to the query, in the POST
body and somehow map it to
query parameters,
the BatchResultsSetPaginator sets the page size to the max size.