add bento result renderer to individual search
This renderer uses the same serializer as before (not optimized),
because the strategy of using a reduced queryset was not playing well with
the search filter.
Indeed, all the logic for filtering (fulltext search)
It is unclear to me where django filter acts in the workflow, and the
smallest change was to keep the existing workflow
and act at the rendering (also last) step.