response consistency: use `subject_id` instead of `id`
This was causing a discrepancy between this response and the one from
table search.
I've tried the other way around (using `id` in both responses) but
it caused an issue due to ambiguous field names at the SQL level.
There is also an added advantage at using a more descriptive name for
the id as the response combines different models.