fix: drf spectacular not served from bento
This is due to Katsu being served from a sub-path,
i.e. not from the portal root.
drf-spectacular was querying at the wrong path.
This relies on an env variable.
Not that SCRIPT_NAME should be "obeyed" by Django, but
I haven't been able to make that
work without using `FORCE_SCRIPT_NAME`
in my tests.